Move WordGroupMatcher + WordParser to the domain layer
