+++ app/src/test/kotlin/de/ywegel/svenska/fakes/WordParserFake.kt
@@ -1,7 +1,7 @@
 package de.ywegel.svenska.fakes
 
-import de.ywegel.svenska.data.WordParser
 import de.ywegel.svenska.data.model.Vocabulary
+import de.ywegel.svenska.domain.wordImporter.WordParser
 
 class WordParserFake : WordParser {
     override fun parseWord(word: String?, translation: String?, containerId: Int): Vocabulary = Vocabulary(
